# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TaskFi API",
    "version": "1.0.0",
    "description": "Web3 Freelancing Platform API - Decentralized marketplace on Solana"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/categories": {
      "get": {
        "summary": "Get all Web3 categories",
        "tags": ["Categories"],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Category" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "Get jobs with pagination",
        "tags": ["Jobs"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "string", "default": "1" } },
          { "name": "limit", "in": "query", "schema": { "type": "string", "default": "10" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of jobs" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create new job",
        "tags": ["Jobs"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateJob" }
            }
          }
        },
        "responses": {
          "201": { "description": "Job created" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/gigs": {
      "get": {
        "summary": "Get gigs with pagination",
        "tags": ["Gigs"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "string", "default": "1" } },
          { "name": "limit", "in": "query", "schema": { "type": "string", "default": "12" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of gigs" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "icon": { "type": "string" },
          "isActive": { "type": "boolean" }
        }
      },
      "CreateJob": {
        "type": "object",
        "required": ["title", "description", "requirements", "budget", "categoryId"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "minLength": 10, "maxLength": 5000 },
          "requirements": { "type": "array", "items": { "type": "string" } },
          "budget": { "type": "number", "minimum": 1, "maximum": 1000000 },
          "categoryId": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time" },
          "isUrgent": { "type": "boolean", "default": false }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}